Problem: Neopixels wrong footprint
Solution: Change to the narrow footprint and place the parts 0.381
inches from each other. This is greater than the assumed 8mm base
across the bottom of the parts we chose, but it gives greater peace
of mind that it will fit. Meanwhile this new footprint has the right
pin pitch that matches the actual diodes, so no spreading pins!